Added support for missing GLFW.h

+++ src/Interface/Interface.cpp
@@ -13,7 +13,7 @@ void Interface::DatabaseWindow() {
 
     // Table names
     if (ImGui::BeginMenuBar()) {
-        for (int i = 0; i < db->getDatabaseData().size(); i++) {
+        for (int i = 0; i < (int)db->getDatabaseData().size(); i++) {
             if (ImGui::MenuItem(&(db->getDatabaseData()[i].getTableName()[0]))) {
                 if (selectedTable != i) {
                     selectedTable = i;
@@ -30,18 +30,18 @@ void Interface::DatabaseWindow() {
         ImGui::Dummy(ImVec2(0, 20 * (int(table.getTableData().size()))));
         int w = (ImGui::GetIO().DisplaySize.x * 0.5 - 38) / std::min(10, int(table.getTableData()[0].getRowData().size())); // Soft limit on 10 visible cols
         ImGui::GetWindowDrawList()->AddRectFilled(ImVec2(cursor.x, cursor.y), ImVec2(cursor.x + 20 - 1, cursor.y + 20 - 1), IM_COL32(30, 30, 30, 255));
-        for (int j = 1; j < table.getTableData().size(); j++) {
+        for (int j = 1; j < (int)table.getTableData().size(); j++) {
             std::string text = std::to_string(j);
             ImGui::GetWindowDrawList()->AddRectFilled(ImVec2(cursor.x, cursor.y + 20*j), ImVec2(cursor.x + 20 - 1, cursor.y + 20*(j + 1) - 1), IM_COL32(30, 30, 30, 255));
             ImGui::GetWindowDrawList()->AddText(ImVec2(((text.length() == 1) ? 5 : 1) + cursor.x, cursor.y + 20*j), IM_COL32(255, 255, 255, 255), &(text[0]), &(text[text.size()]));
         }
-        for (int j = 0; j < table.getTableData()[0].getRowData().size(); j++) {
+        for (int j = 0; j < (int)table.getTableData()[0].getRowData().size(); j++) {
             std::string text = table.getTableData()[0].getRowData()[j].getValue();
             ImGui::GetWindowDrawList()->AddRectFilled(ImVec2(20 + cursor.x + w*j, cursor.y), ImVec2(20 + cursor.x + w*(j + 1) - 1, cursor.y + 20 - 1), IM_COL32(40, 50, 30, 255));
             ImGui::GetWindowDrawList()->AddText(ImVec2(20 + 5 + cursor.x + w*j, cursor.y), IM_COL32(255, 255, 255, 255), &(text[0]), &(text[text.size()]));
         }
-        for (int i = 1; i < table.getTableData().size(); i++) {
-            for (int j = 0; j < table.getTableData()[i].getRowData().size(); j++) {
+        for (int i = 1; i < (int)table.getTableData().size(); i++) {
+            for (int j = 0; j < (int)table.getTableData()[i].getRowData().size(); j++) {
                 std::string text = table.getTableData()[i].getRowData()[j].getValue();
                 ImGui::GetWindowDrawList()->AddRectFilled(ImVec2(20 + cursor.x + w*j, cursor.y + 20*i), ImVec2(20 + cursor.x + w*(j + 1) - 1, cursor.y + 20*(i + 1) - 1), IM_COL32(30, 30, 30, 255));
                 if (ImGui::IsMouseHoveringRect(ImVec2(20 + cursor.x + w*j, cursor.y + 20*i), ImVec2(20 + cursor.x + w*(j + 1) - 1, cursor.y + 20*(i + 1) - 1))) {
@@ -74,7 +74,7 @@ void Interface::EditorWindow(int buttonHeight) {
 
     // Syntax Highlight
     std::vector<std::string> wordList = lexer->getSpacedWords(queryBuffer);
-    for (int i = 0; i < wordList.size(); i++) {
+    for (int i = 0; i < (int)wordList.size(); i++) {
         if (lexer->lower(lexer->strip(wordList[i])) == "create") {
             ImGui::GetWindowDrawList()->AddText(ImVec2(ImGui::GetCursorScreenPos().x + 3, ImGui::GetCursorScreenPos().y + 3), IM_COL32(50, 150, 200, 255), &(wordList[i][0]));
         } else if (lexer->lower(lexer->strip(wordList[i])) == "addrow") {
@@ -154,18 +154,18 @@ void Interface::EditorResponse() {
         Table& table = *((*READRESPONSE)[0]);
         int w = (ImGui::GetIO().DisplaySize.x * 0.5 - 38) / std::min(10, int(table.getTableData()[0].getRowData().size())); // Soft limit on 10 visible cols
         ImGui::GetWindowDrawList()->AddRectFilled(ImVec2(cursor.x, cursor.y), ImVec2(cursor.x + 20 - 1, cursor.y + 20 - 1), IM_COL32(30, 30, 30, 255));
-        for (int j = 1; j < table.getTableData().size(); j++) {
+        for (int j = 1; j < (int)table.getTableData().size(); j++) {
             std::string text = std::to_string(j);
             ImGui::GetWindowDrawList()->AddRectFilled(ImVec2(cursor.x, cursor.y + 20*j), ImVec2(cursor.x + 20 - 1, cursor.y + 20*(j + 1) - 1), IM_COL32(30, 30, 30, 255));
             ImGui::GetWindowDrawList()->AddText(ImVec2(((text.length() == 1) ? 5 : 1) + cursor.x, cursor.y + 20*j), IM_COL32(255, 255, 255, 255), &(text[0]), &(text[text.size()]));
         }
-        for (int j = 0; j < table.getTableData()[0].getRowData().size(); j++) {
+        for (int j = 0; j < (int)table.getTableData()[0].getRowData().size(); j++) {
             std::string text = table.getTableData()[0].getRowData()[j].getValue();
             ImGui::GetWindowDrawList()->AddRectFilled(ImVec2(20 + cursor.x + w*j, cursor.y), ImVec2(20 + cursor.x + w*(j + 1) - 1, cursor.y + 20 - 1), IM_COL32(40, 50, 30, 255));
             ImGui::GetWindowDrawList()->AddText(ImVec2(20 + 5 + cursor.x + w*j, cursor.y), IM_COL32(255, 255, 255, 255), &(text[0]), &(text[text.size()]));
         }
-        for (int i = 1; i < table.getTableData().size(); i++) {
-            for (int j = 0; j < table.getTableData()[i].getRowData().size(); j++) {
+        for (int i = 1; i < (int)table.getTableData().size(); i++) {
+            for (int j = 0; j < (int)table.getTableData()[i].getRowData().size(); j++) {
                 std::string text = table.getTableData()[i].getRowData()[j].getValue();
                 ImGui::GetWindowDrawList()->AddRectFilled(ImVec2(20 + cursor.x + w*j, cursor.y + 20*i), ImVec2(20 + cursor.x + w*(j + 1) - 1, cursor.y + 20*(i + 1) - 1), IM_COL32(30, 30, 30, 255));
                 if (ImGui::IsMouseHoveringRect(ImVec2(20 + cursor.x + w*j, cursor.y + 20*i), ImVec2(20 + cursor.x + w*(j + 1) - 1, cursor.y + 20*(i + 1) - 1))) {
